refactor: modularize app sections
